bugfix: exit with no exception specified when no PVT signal (some GPGGA data fields empty))
When running the program with no PVT signal, it would crash as soon as GPGGA message was received. The reason for that is the NMEAToDecimal function that is called when some of the fields in the GPGGA message are empty (containing the longitude and latitude information). Now, the library can handle this scenario without crashing.